ummm a whole buncha elements shit

+++ pets-lib/src/battle/skills/status_effects.rs
@@ -49,6 +49,6 @@ impl EffectChance {
             Rare => EFFECT_CHANCE_RARE,
         };
 
-        rand::random::<f32>() < chance
+        rand::random::<f64>() < chance
     }
 }
